fix: Project Restructure
This commit aims to fix all the problems caused by PR#8, along with some
fixes
- Remove package-lock.json as we're already using yarn.package-lock
- Remove src/ folder
- Fix paths in various files
